requirements.txt
firebase-admin==6.8.0
google-adk==1.1.1